requirements.txt
appdirs==1.4.3
Django==1.10.4
django-autoslug==1.9.3
django-mptt==0.8.7
django-widget-tweaks==1.4.1
Faker==0.7.3
ipaddress==1.0.18
mixer==5.6.6
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytils==0.3
pytz==2017.2
six==1.10.0
